+++ static/js/addons/cl.utils.js
@@ -8,6 +8,7 @@
 /**
  * @module Cl
  */
+// istanbul ignore next
 var Cl = window.Cl || {};
 
 // #####################################################################################################################
@@ -44,6 +45,7 @@ var Cl = window.Cl || {};
          */
         _consoleWrapper: function () {
             var method;
+            // istanbul ignore next
             var noop = function () {};
             var methods = [
                 'assert', 'clear', 'count', 'debug', 'dir', 'dirxml', 'error',
@@ -52,6 +54,7 @@ var Cl = window.Cl || {};
                 'timeStamp', 'trace', 'warn'
             ];
             var length = methods.length;
+            // istanbul ignore next
             var console = (window.console = window.console || {});
 
             while (length--) {
@@ -82,9 +85,11 @@ var Cl = window.Cl || {};
          * @param value {String} storage value
          */
         setStorage: function (token, value) {
-            if (typeof (Storage) !== void(0)) {
+            if (token && value && this._isStorageSupported) {
                 localStorage.setItem(token, value);
                 return value;
+            } else {
+                return false;
             }
         },
 
@@ -95,10 +100,29 @@ var Cl = window.Cl || {};
          * @param token {String} namespace
          */
         getStorage: function (token) {
-            if (typeof (Storage) !== void(0)) {
+            if (token && this._isStorageSupported) {
                 return localStorage.getItem(token);
+            } else {
+                return false;
             }
-        }
+        },
+
+        /**
+         * Localstorage shim from Modernizr
+         *
+         * @method _isStorageSupported
+         * @private
+         */
+        _isStorageSupported: (function localStorageCheck() {
+            var mod = 'modernizr';
+            try {
+                localStorage.setItem(mod, mod);
+                localStorage.removeItem(mod);
+                return true;
+            } catch (e) {
+                return false;
+            }
+        }())
     };
 
 })(jQuery);